requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
cloudinary==1.32.0
crispy-bootstrap4==2022.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.18
django-allauth==0.51.0
django-bootstrap-datepicker-plus==5.0.3
django-crispy-forms==2.0
django-phonenumber-field==7.0.2
django-summernote==0.8.20.0
gunicorn==20.1.0
oauthlib==3.2.2
pathspec==0.11.1
phonenumbers==8.13.7
psycopg2==2.9.5
pydantic==1.10.7
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
